chore: move keychain to libp2p
