--- .gitattributes
@@ -1,2 +1,0 @@
-*.png binary
-*	crlf=input
