+++ package.json
@@ -49,6 +49,7 @@
     "debug": "^4.1.1",
     "err-code": "^1.1.2",
     "hashlru": "^2.3.0",
+    "interface-datastore": "^0.8.0",
     "it-all": "^1.0.1",
     "it-buffer": "^0.1.1",
     "it-handshake": "^1.0.1",
@@ -64,6 +65,7 @@
     "multiaddr": "^7.2.1",
     "multistream-select": "^0.15.0",
     "mutable-proxy": "^1.0.0",
+    "node-forge": "^0.9.1",
     "p-any": "^2.1.0",
     "p-fifo": "^1.0.0",
     "p-settle": "^3.1.0",
@@ -71,6 +73,7 @@
     "peer-info": "^0.17.0",
     "protons": "^1.0.1",
     "retimer": "^2.0.0",
+    "sanitize-filename": "^1.6.3",
     "timeout-abort-controller": "^1.0.0",
     "xsalsa20": "^1.0.2"
   },
@@ -80,12 +83,17 @@
     "aegir": "^20.5.1",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
+    "chai-string": "^1.5.0",
     "cids": "^0.7.1",
+    "datastore-fs": "^0.9.1",
+    "datastore-level": "^0.14.0",
     "delay": "^4.3.0",
     "dirty-chai": "^2.0.1",
+    "is-browser": "^2.1.0",
     "it-concat": "^1.0.0",
     "it-pair": "^1.0.0",
     "it-pushable": "^1.4.0",
+    "level": "^6.0.0",
     "libp2p-bootstrap": "^0.10.3",
     "libp2p-delegated-content-routing": "^0.4.1",
     "libp2p-delegated-peer-routing": "^0.4.0",
@@ -98,10 +106,13 @@
     "libp2p-tcp": "^0.14.1",
     "libp2p-webrtc-star": "^0.17.0",
     "libp2p-websockets": "^0.13.1",
+    "multihashes": "^0.4.15",
     "nock": "^10.0.6",
     "p-defer": "^3.0.0",
     "p-times": "^2.1.0",
     "p-wait-for": "^3.1.0",
+    "promisify-es6": "^1.0.3",
+    "rimraf": "^3.0.0",
     "sinon": "^8.1.0",
     "streaming-iterables": "^4.1.0",
     "wrtc": "^0.4.1"
