+++ test/keychain/peerid.spec.js
